+++ NEWS.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.23] unreleased
+
+### Added
+
+* `HermitianPositiveDefinite` manifold of Hermitian positive definite matrices.
+* `MatrixSqrtManifoldPoint` as a generalization/replacement for the `SPDPoint`
+
 ## [0.10.22] 2025-06-25
 
 ### Fixed